Created routes directory and booksRoute.js file, transferred all api routes from index.js to booksRoute.js file
