+++ backend/index.js
@@ -1,7 +1,7 @@
 import express from "express";
 import { PORT, mongoDBURL } from "./config.js";
 import mongoose from "mongoose";
-import { Book } from "./models/bookModel.js";
+import booksRoute from "./routes/booksRoute.js";
 
 const app = express();
 
@@ -15,108 +15,8 @@ app.get("/", (request, response) => {
     .send("Welcome to your first API endpoint in nodejs");
 });
 
-// Route for Save/POST a new Book
-
-app.post("/books", async (request, response) => {
-  try {
-    if (
-      !request.body.title ||
-      !request.body.author ||
-      !request.body.publishYear
-    ) {
-      return response.status(400).send({
-        message: "Title, Author, and Publish year fields are required.",
-      });
-    }
-    const newBook = {
-      title: request.body.title,
-      author: request.body.author,
-      publishYear: request.body.publishYear,
-    };
-
-    const book = await Book.create(newBook);
-
-    return response.status(201).send(book);
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send(error.message);
-  }
-});
-
-// Route to GET all books from database
-app.get("/books", async (request, response) => {
-  try {
-    const books = await Book.find({});
-
-    return response.status(200).json({
-      count: books.length,
-      data: books,
-    });
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send(error.message);
-  }
-});
-
-// Route to GET a book from database by id
-app.get("/books/:id", async (request, response) => {
-  try {
-    const { id } = request.params;
-
-    const book = await Book.findById(id);
-
-    return response.status(200).json(book);
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-// Route to update a book
-app.put("/books/:id", async (request, response) => {
-  try {
-    if (
-      !request.body.title ||
-      !request.body.author ||
-      !request.body.publishYear
-    ) {
-      return response.status(400).send({
-        message: "Title, Author, and Publish year fields are required.",
-      });
-    }
-
-    const { id } = request.params;
-
-    const result = await Book.findByIdAndUpdate(id, request.body);
-
-    if (!result) {
-      return response.status(404).json({ message: "Book not found" });
-    }
-
-    return response.status(200).send({ message: "Book update was successful" });
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-// Route to delete a book
-app.delete("/books/:id", async (request, response) => {
-  try {
-    const { id } = request.params;
-
-    const result = await Book.findByIdAndDelete(id);
-
-    if(!result){
-      return  response.status(404).json({message: 'Book not found'});
-    }
-
-    return response.status(200).send({message: 'Book deleted successfully'});
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
+// Middleware for books api routes
+app.use("/books", booksRoute);
 
 mongoose
   .connect(mongoDBURL)
